// source code/front-end-user/src/app/models/user-role.enum.ts
export enum UserRole {
  // ROLE_ADMIN = 'ROLE_ADMIN',
  // ROLE_STUDENT = 'ROLE_STUDENT',
  // ROLE_LECTURE = 'ROLE_LECTURER',
  ROLE_CEO = 'ROLE_CEO',
  ROLE_BRANCH_MANAGER = 'ROLE_BRANCH_MANAGER',
  ROLE_TRANSACTION_CLERK = 'ROLE_TRANSACTION_CLERK',
  ROLE_WAREHOUSE_MANAGER = 'ROLE_WAREHOUSE_MANAGER',
  ROLE_WAREHOUSE_CLERK = 'ROLE_WAREHOUSE_CLERK',
  ROLE_USER = 'ROLE_USER'
}